Daemonizes rsense. Fixes #11
When `bin/rsense start` is called from the command line it uses the
`jruby-jarsgem` to get the location of the `jruby-complete` jar. Next
it writes the current Ruby's gem information into the ENV, so it isn't
 lost during the "fork". It spawns a daemonized java process, via the
 `Spoon` gem, that launches jruby, placing rsense on the load path, and
 requiring `bin/_rsense.rb` which is now running on JRuby, but has
access to the host ruby's gems. This last is critical, as we would
otherwise have to somehow bundle all the gems and files rsense needs,
including rsense-core and jruby-parser, into one gem, which would
likely be to big and rubygems.org. The spawned process can be stopped
by calling `bin/rsense stop`.