Client now locates Daemon/Server code in `rsense-server` gem.
As previously described, this is a neccessary architectural change.
Besides obvious benefits of modularity, clearer delineation of concerns,
etc, this was needed for testability. It was previously impossible to
test any of the code requiring java without either losing the ability
to test the client code with MRI, or doing weird, complicated stuff.